refactor: split cleanup integration evaluators
Extract Seerr, Plex, and Jellyfin predicate families behind explicit pure adapter modules while retaining central filter policy, composite semantics, and dispatch. Shared normalized lookups are exported instead of duplicated. All 2,241 API tests pass, API typecheck passes, and Biome reports only existing baseline warnings.
