+++ apps/api/src/lib/library-cleanup/rule-evaluators.ts
@@ -63,15 +63,42 @@ import { safeJsonParse } from "../utils/json.js";
 import type {
 	CacheItemForEval,
 	EvalContext,
-	JellyfinWatchInfo,
-	JellyfinWatchMap,
-	PlexWatchInfo,
 	PlexWatchMap,
 	RuleAction,
 	RuleMatch,
-	SeerrRequestInfo,
 	SeerrRequestMap,
 } from "./types.js";
+import {
+	evaluateJellyfinAddedAt,
+	evaluateJellyfinEpisodeCompletion,
+	evaluateJellyfinLastWatched,
+	evaluateJellyfinOnDeck,
+	evaluateJellyfinUserRating,
+	evaluateJellyfinWatchCount,
+	evaluateJellyfinWatchedBy,
+} from "./evaluators/jellyfin-evaluators.js";
+import {
+	evaluatePlexAddedAt,
+	evaluatePlexCollection,
+	evaluatePlexLabel,
+	evaluatePlexLastWatched,
+	evaluatePlexOnDeck,
+	evaluatePlexUserRating,
+	evaluatePlexWatchCount,
+	evaluatePlexWatchedBy,
+	lookupPlexWatch,
+} from "./evaluators/plex-evaluators.js";
+import {
+	evaluateSeerrIs4k,
+	evaluateSeerrIsRequested,
+	evaluateSeerrModifiedBy,
+	evaluateSeerrRequestAge,
+	evaluateSeerrRequestCount,
+	evaluateSeerrRequestedBy,
+	evaluateSeerrRequestModifiedAge,
+	evaluateSeerrRequestStatus,
+	lookupSeerrRequests,
+} from "./evaluators/seerr-evaluators.js";
 
 // ============================================================================
 // Per-Type Evaluators
@@ -519,238 +546,6 @@ function evaluateReleaseGroupRule(
 }
 
 // ============================================================================
-// Seerr Rule Evaluators
-// ============================================================================
-
-/** Seerr request status code → label mapping */
-const SEERR_STATUS_LABELS: Record<number, string> = {
-	1: "pending",
-	2: "approved",
-	3: "declined",
-	4: "failed",
-	5: "completed",
-};
-
-/**
- * Look up Seerr requests for a cache item via its tmdbId from the data blob.
- */
-function lookupSeerrRequests(
-	item: CacheItemForEval,
-	seerrMap: SeerrRequestMap | undefined,
-): SeerrRequestInfo[] | null {
-	if (!seerrMap || seerrMap.size === 0) return null;
-
-	const parsed = safeJsonParse(item.data);
-	if (!parsed) return null;
-	const data = parsed as Record<string, unknown>;
-
-	// Extract tmdbId from the data blob (LibraryItem.remoteIds.tmdbId)
-	const remoteIds = data.remoteIds as Record<string, unknown> | undefined;
-	const tmdbId = remoteIds?.tmdbId;
-	if (!tmdbId) return null;
-
-	// Build the lookup key: "movie:12345" or "tv:12345"
-	const mediaType = item.itemType === "movie" ? "movie" : "tv";
-	const key = `${mediaType}:${tmdbId}`;
-
-	return seerrMap.get(key) ?? null;
-}
-
-/**
- * Seerr Requested By: flag items requested by specific Seerr users.
- */
-function evaluateSeerrRequestedBy(
-	item: CacheItemForEval,
-	params: SeerrRequestedByParams,
-	seerrMap: SeerrRequestMap | undefined,
-): string | null {
-	const requests = lookupSeerrRequests(item, seerrMap);
-	if (!requests || requests.length === 0) return null;
-
-	const targetNames = params.userNames.map((n) => n.toLowerCase());
-
-	for (const req of requests) {
-		if (targetNames.includes(req.requestedBy.toLowerCase())) {
-			return `Requested by "${req.requestedBy}" (match: [${params.userNames.join(", ")}])`;
-		}
-	}
-
-	return null;
-}
-
-/**
- * Seerr Request Age: flag items whose Seerr request is older/newer than N days.
- */
-function evaluateSeerrRequestAge(
-	item: CacheItemForEval,
-	params: SeerrRequestAgeParams,
-	seerrMap: SeerrRequestMap | undefined,
-	now: Date,
-): string | null {
-	const requests = lookupSeerrRequests(item, seerrMap);
-	if (!requests || requests.length === 0) return null;
-
-	// Use the oldest request's createdAt
-	let oldest = requests[0]!;
-	for (const req of requests) {
-		if (req.createdAt < oldest.createdAt) oldest = req;
-	}
-
-	const requestDate = new Date(oldest.createdAt);
-	const ageDays = (now.getTime() - requestDate.getTime()) / (1000 * 60 * 60 * 24);
-
-	if (params.operator === "older_than" && ageDays >= params.days) {
-		return `Seerr request ${Math.floor(ageDays)} days old (threshold: > ${params.days} days, requested by ${oldest.requestedBy})`;
-	}
-	if (params.operator === "newer_than" && ageDays < params.days) {
-		return `Seerr request ${Math.floor(ageDays)} days old (threshold: < ${params.days} days, requested by ${oldest.requestedBy})`;
-	}
-
-	return null;
-}
-
-/**
- * Seerr Request Status: flag items whose Seerr request has a specific status.
- */
-function evaluateSeerrRequestStatus(
-	item: CacheItemForEval,
-	params: SeerrRequestStatusParams,
-	seerrMap: SeerrRequestMap | undefined,
-): string | null {
-	const requests = lookupSeerrRequests(item, seerrMap);
-	if (!requests || requests.length === 0) return null;
-
-	for (const req of requests) {
-		const statusLabel = SEERR_STATUS_LABELS[req.status];
-		if (statusLabel && params.statuses.includes(statusLabel as (typeof params.statuses)[number])) {
-			return `Seerr request status is "${statusLabel}" (requested by ${req.requestedBy})`;
-		}
-	}
-
-	return null;
-}
-
-/**
- * Seerr Is 4K: flag items based on whether the Seerr request is for 4K.
- */
-function evaluateSeerrIs4k(
-	item: CacheItemForEval,
-	params: SeerrIs4kParams,
-	seerrMap: SeerrRequestMap | undefined,
-): string | null {
-	const requests = lookupSeerrRequests(item, seerrMap);
-	if (!requests || requests.length === 0) return null;
-
-	for (const req of requests) {
-		if (req.is4k === params.is4k) {
-			return params.is4k
-				? `Seerr request is 4K (requested by ${req.requestedBy})`
-				: `Seerr request is not 4K (requested by ${req.requestedBy})`;
-		}
-	}
-	return null;
-}
-
-/**
- * Seerr Request Modified Age: flag items by how recently the Seerr request was modified.
- */
-function evaluateSeerrRequestModifiedAge(
-	item: CacheItemForEval,
-	params: SeerrRequestModifiedAgeParams,
-	seerrMap: SeerrRequestMap | undefined,
-	now: Date,
-): string | null {
-	const requests = lookupSeerrRequests(item, seerrMap);
-	if (!requests || requests.length === 0) return null;
-
-	// Use the most recently modified request
-	let latest = requests[0]!;
-	for (const req of requests) {
-		if (req.updatedAt && req.updatedAt > (latest.updatedAt ?? "")) latest = req;
-	}
-
-	if (!latest.updatedAt) return null;
-	const modifiedDate = new Date(latest.updatedAt);
-	const ageDays = (now.getTime() - modifiedDate.getTime()) / (1000 * 60 * 60 * 24);
-
-	if (params.operator === "older_than" && ageDays >= params.days) {
-		return `Seerr request last modified ${Math.floor(ageDays)} days ago (threshold: > ${params.days} days)`;
-	}
-	if (params.operator === "newer_than" && ageDays < params.days) {
-		return `Seerr request last modified ${Math.floor(ageDays)} days ago (threshold: < ${params.days} days)`;
-	}
-	return null;
-}
-
-/**
- * Seerr Modified By: flag items whose Seerr request was last modified by specific users.
- */
-function evaluateSeerrModifiedBy(
-	item: CacheItemForEval,
-	params: SeerrModifiedByParams,
-	seerrMap: SeerrRequestMap | undefined,
-): string | null {
-	const requests = lookupSeerrRequests(item, seerrMap);
-	if (!requests || requests.length === 0) return null;
-
-	const targetNames = params.userNames.map((n) => n.toLowerCase());
-
-	for (const req of requests) {
-		if (req.modifiedBy && targetNames.includes(req.modifiedBy.toLowerCase())) {
-			return `Seerr request modified by "${req.modifiedBy}" (match: [${params.userNames.join(", ")}])`;
-		}
-	}
-	return null;
-}
-
-/**
- * Seerr Is Requested: flag items based on whether they have any Seerr request.
- */
-function evaluateSeerrIsRequested(
-	item: CacheItemForEval,
-	params: SeerrIsRequestedParams,
-	seerrMap: SeerrRequestMap | undefined,
-): string | null {
-	// If Seerr data is unavailable, skip evaluation to avoid false "not requested" matches
-	if (!seerrMap) return null;
-	const requests = lookupSeerrRequests(item, seerrMap);
-	const hasRequest = requests !== null && requests.length > 0;
-
-	if (params.isRequested && hasRequest) {
-		return `Has Seerr request (${requests!.length} request(s))`;
-	}
-	if (!params.isRequested && !hasRequest) {
-		return "No Seerr request found";
-	}
-	return null;
-}
-
-/**
- * Seerr Request Count: flag items based on number of Seerr requests.
- */
-function evaluateSeerrRequestCount(
-	item: CacheItemForEval,
-	params: SeerrRequestCountParams,
-	seerrMap: SeerrRequestMap | undefined,
-): string | null {
-	if (!seerrMap) return null;
-	const requests = lookupSeerrRequests(item, seerrMap);
-	// If lookup returns null (no tmdbId), skip to avoid false "0 requests" matches
-	if (requests === null) return null;
-	const count = requests.length;
-
-	if (params.operator === "less_than" && count < params.count) {
-		return `Seerr request count: ${count} (threshold: < ${params.count})`;
-	}
-	if (params.operator === "greater_than" && count > params.count) {
-		return `Seerr request count: ${count} (threshold: > ${params.count})`;
-	}
-	if (params.operator === "equals" && count === params.count) {
-		return `Seerr request count: ${count} (threshold: = ${params.count})`;
-	}
-	return null;
-}
-
 // ============================================================================
 // Requester-Aware Cross-Service Evaluators
 // ============================================================================
@@ -820,522 +615,7 @@ function evaluateSeerrRequesterNotWatched(
 }
 
 // ============================================================================
-// Plex Rule Evaluators
 // ============================================================================
-
-/**
- * Look up Plex watch data for a cache item via its tmdbId from the data blob.
- * When plexLibraryFilter is set, only data from matching Plex library sections is used.
- */
-function lookupPlexWatch(
-	item: CacheItemForEval,
-	plexMap: PlexWatchMap | undefined,
-	plexLibraryFilter?: string[] | null,
-): PlexWatchInfo | null {
-	if (!plexMap || plexMap.size === 0) return null;
-
-	const parsed = safeJsonParse(item.data);
-	if (!parsed) return null;
-	const data = parsed as Record<string, unknown>;
-
-	const remoteIds = data.remoteIds as Record<string, unknown> | undefined;
-	const tmdbId = remoteIds?.tmdbId;
-	if (!tmdbId) return null;
-
-	const mediaType = item.itemType === "movie" ? "movie" : "series";
-	const key = `${mediaType}:${tmdbId}`;
-
-	const entry = plexMap.get(key);
-	if (!entry) return null;
-
-	// If no filter, return the pre-computed aggregates (existing behavior)
-	if (!plexLibraryFilter || plexLibraryFilter.length === 0) return entry;
-
-	// Filter to matching sections only
-	const matchingSections = entry.sections.filter((s) => plexLibraryFilter.includes(s.sectionTitle));
-	if (matchingSections.length === 0) return null;
-
-	// Re-aggregate from filtered sections
-	return {
-		lastWatchedAt: matchingSections.reduce<Date | null>((latest, s) => {
-			if (!s.lastWatchedAt) return latest;
-			return !latest || s.lastWatchedAt > latest ? s.lastWatchedAt : latest;
-		}, null),
-		watchCount: matchingSections.reduce((sum, s) => sum + s.watchCount, 0),
-		watchedByUsers: [...new Set(matchingSections.flatMap((s) => s.watchedByUsers))],
-		onDeck: matchingSections.some((s) => s.onDeck),
-		userRating: matchingSections.reduce<number | null>((best, s) => {
-			if (s.userRating == null) return best;
-			return best != null ? Math.max(best, s.userRating) : s.userRating;
-		}, null),
-		collections: [...new Set(matchingSections.flatMap((s) => s.collections))],
-		labels: [...new Set(matchingSections.flatMap((s) => s.labels))],
-		addedAt: matchingSections.reduce<Date | null>((earliest, s) => {
-			if (!s.addedAt) return earliest;
-			return !earliest || s.addedAt < earliest ? s.addedAt : earliest;
-		}, null),
-		sections: matchingSections,
-	};
-}
-
-/**
- * Plex Last Watched: flag items based on when they were last watched.
- * "never" operator flags items that have never been watched.
- * "older_than" uses addedAt as a fallback for never-watched items — if an item
- * was added N+ days ago and never watched, it qualifies as "unwatched for N+ days".
- */
-function evaluatePlexLastWatched(
-	item: CacheItemForEval,
-	params: PlexLastWatchedParams,
-	ctx: EvalContext,
-	plexLibraryFilter?: string[] | null,
-): string | null {
-	const watch = lookupPlexWatch(item, ctx.plexMap, plexLibraryFilter);
-
-	if (params.operator === "never") {
-		if (!watch || watch.lastWatchedAt === null) {
-			return "Never watched (per Plex)";
-		}
-		return null;
-	}
-
-	if (!watch) return null;
-
-	if (watch.lastWatchedAt) {
-		const ageDays = (ctx.now.getTime() - watch.lastWatchedAt.getTime()) / (1000 * 60 * 60 * 24);
-		if (params.operator === "older_than" && params.days && ageDays >= params.days) {
-			return `Last watched ${Math.floor(ageDays)} days ago in Plex (threshold: > ${params.days} days)`;
-		}
-		return null;
-	}
-
-	// Never watched — fall back to addedAt (Plex), then arrAddedAt (Sonarr/Radarr)
-	if (params.operator === "older_than" && params.days) {
-		const refDate = watch.addedAt ?? item.arrAddedAt;
-		if (refDate) {
-			const addedDays = (ctx.now.getTime() - refDate.getTime()) / (1000 * 60 * 60 * 24);
-			if (addedDays >= params.days) {
-				const source = watch.addedAt ? "Plex" : "library";
-				return `Never watched, added to ${source} ${Math.floor(addedDays)} days ago (threshold: > ${params.days} days)`;
-			}
-		}
-	}
-
-	return null;
-}
-
-/**
- * Plex Watch Count: flag items based on play count.
- */
-function evaluatePlexWatchCount(
-	item: CacheItemForEval,
-	params: PlexWatchCountParams,
-	ctx: EvalContext,
-	plexLibraryFilter?: string[] | null,
-): string | null {
-	const watch = lookupPlexWatch(item, ctx.plexMap, plexLibraryFilter);
-	if (!watch) {
-		// Not in Plex — infer 0 plays when Plex is configured and item has a file
-		if (
-			ctx.plexMap &&
-			ctx.plexMap.size > 0 &&
-			params.operator === "less_than" &&
-			params.count > 0 &&
-			item.hasFile &&
-			item.arrAddedAt
-		) {
-			const ageDays = Math.floor(
-				(ctx.now.getTime() - item.arrAddedAt.getTime()) / (1000 * 60 * 60 * 24),
-			);
-			return `Not tracked by Plex, in library for ${ageDays} days (threshold: < ${params.count} plays)`;
-		}
-		return null;
-	}
-	const count = watch.watchCount;
-
-	// Age context for low play counts
-	const ageCtx =
-		count === 0 && watch.addedAt
-			? `, added ${Math.floor((ctx.now.getTime() - watch.addedAt.getTime()) / (1000 * 60 * 60 * 24))} days ago`
-			: "";
-
-	if (params.operator === "less_than" && count < params.count) {
-		return `Plex play count: ${count}${ageCtx} (threshold: < ${params.count})`;
-	}
-	if (params.operator === "greater_than" && count > params.count) {
-		return `Plex play count: ${count} (threshold: > ${params.count})`;
-	}
-	return null;
-}
-
-/**
- * Plex On Deck: flag items based on whether they are on Plex's Continue Watching.
- */
-function evaluatePlexOnDeck(
-	item: CacheItemForEval,
-	params: PlexOnDeckParams,
-	ctx: EvalContext,
-	plexLibraryFilter?: string[] | null,
-): string | null {
-	const watch = lookupPlexWatch(item, ctx.plexMap, plexLibraryFilter);
-	if (!watch) return null;
-	const isOnDeck = watch.onDeck;
-
-	if (params.isDeck && isOnDeck) {
-		return "Item is on Plex Continue Watching";
-	}
-	if (!params.isDeck && !isOnDeck) {
-		return "Item is not on Plex Continue Watching";
-	}
-	return null;
-}
-
-/**
- * Plex User Rating: flag items based on the admin's star rating in Plex.
- * "unrated" operator flags items with no rating.
- */
-function evaluatePlexUserRating(
-	item: CacheItemForEval,
-	params: PlexUserRatingParams,
-	ctx: EvalContext,
-	plexLibraryFilter?: string[] | null,
-): string | null {
-	const watch = lookupPlexWatch(item, ctx.plexMap, plexLibraryFilter);
-
-	if (params.operator === "unrated") {
-		if (!watch || watch.userRating === null) {
-			return "Unrated in Plex";
-		}
-		return null;
-	}
-
-	if (!watch || watch.userRating === null) return null;
-
-	if (
-		params.operator === "less_than" &&
-		params.rating !== undefined &&
-		watch.userRating < params.rating
-	) {
-		return `Plex user rating: ${watch.userRating.toFixed(1)} (threshold: < ${params.rating})`;
-	}
-	if (
-		params.operator === "greater_than" &&
-		params.rating !== undefined &&
-		watch.userRating > params.rating
-	) {
-		return `Plex user rating: ${watch.userRating.toFixed(1)} (threshold: > ${params.rating})`;
-	}
-	return null;
-}
-
-/**
- * Plex Watched By: flag items based on which Plex users have watched them.
- */
-function evaluatePlexWatchedBy(
-	item: CacheItemForEval,
-	params: PlexWatchedByParams,
-	ctx: EvalContext,
-	plexLibraryFilter?: string[] | null,
-): string | null {
-	const watch = lookupPlexWatch(item, ctx.plexMap, plexLibraryFilter);
-	if (!watch) return null;
-	const watchedBy = watch.watchedByUsers.map((u) => u.toLowerCase());
-	const targetNames = params.userNames.map((n) => n.toLowerCase());
-
-	if (params.operator === "includes_any") {
-		const matched = targetNames.filter((n) => watchedBy.includes(n));
-		if (matched.length > 0) {
-			return `Watched by Plex user(s): ${matched.join(", ")}`;
-		}
-	} else if (params.operator === "excludes_all") {
-		const noneWatched = targetNames.every((n) => !watchedBy.includes(n));
-		if (noneWatched) {
-			return `Not watched by Plex user(s): ${params.userNames.join(", ")}`;
-		}
-	}
-
-	return null;
-}
-
-/**
- * Plex Collection: flag items based on Plex collection membership.
- */
-function evaluatePlexCollection(
-	item: CacheItemForEval,
-	params: PlexCollectionRuleParams,
-	ctx: EvalContext,
-	plexLibraryFilter?: string[] | null,
-): string | null {
-	const watch = lookupPlexWatch(item, ctx.plexMap, plexLibraryFilter);
-	if (!watch) return null;
-	const collections = watch.collections;
-
-	const targetLower = params.collections.map((c) => c.toLowerCase());
-	const itemLower = collections.map((c) => c.toLowerCase());
-
-	if (params.operator === "in") {
-		const matched = targetLower.filter((c) => itemLower.includes(c));
-		if (matched.length > 0) {
-			return `In Plex collection(s): ${matched.join(", ")}`;
-		}
-	} else if (params.operator === "not_in") {
-		const hasNone = targetLower.every((c) => !itemLower.includes(c));
-		if (hasNone) {
-			return `Not in Plex collection(s): ${params.collections.join(", ")}`;
-		}
-	}
-	return null;
-}
-
-/**
- * Plex Label: flag items based on Plex label tags.
- */
-function evaluatePlexLabel(
-	item: CacheItemForEval,
-	params: PlexLabelRuleParams,
-	ctx: EvalContext,
-	plexLibraryFilter?: string[] | null,
-): string | null {
-	const watch = lookupPlexWatch(item, ctx.plexMap, plexLibraryFilter);
-	if (!watch) return null;
-	const labels = watch.labels;
-
-	const targetLower = params.labels.map((l) => l.toLowerCase());
-	const itemLower = labels.map((l) => l.toLowerCase());
-
-	if (params.operator === "has_any") {
-		const matched = targetLower.filter((l) => itemLower.includes(l));
-		if (matched.length > 0) {
-			return `Has Plex label(s): ${matched.join(", ")}`;
-		}
-	} else if (params.operator === "has_none") {
-		const hasNone = targetLower.every((l) => !itemLower.includes(l));
-		if (hasNone) {
-			return `Does not have Plex label(s): ${params.labels.join(", ")}`;
-		}
-	}
-	return null;
-}
-
-/**
- * Plex Added At: flag items based on when they were added to Plex.
- * "older_than" flags items added more than N days ago.
- * "newer_than" flags items added less than N days ago.
- */
-function evaluatePlexAddedAt(
-	item: CacheItemForEval,
-	params: PlexAddedAtParams,
-	ctx: EvalContext,
-	plexLibraryFilter?: string[] | null,
-): string | null {
-	const watch = lookupPlexWatch(item, ctx.plexMap, plexLibraryFilter);
-	if (!watch?.addedAt) return null;
-
-	const ageDays = (ctx.now.getTime() - watch.addedAt.getTime()) / (1000 * 60 * 60 * 24);
-
-	if (params.operator === "older_than" && ageDays >= params.days) {
-		return `Added to Plex ${Math.floor(ageDays)} days ago (threshold: > ${params.days} days)`;
-	}
-	if (params.operator === "newer_than" && ageDays < params.days) {
-		return `Added to Plex ${Math.floor(ageDays)} days ago (threshold: < ${params.days} days)`;
-	}
-	return null;
-}
-
-// ============================================================================
-// Jellyfin Rule Evaluators
-// ============================================================================
-
-function lookupJellyfinWatch(
-	item: CacheItemForEval,
-	jellyfinMap: JellyfinWatchMap | undefined,
-): JellyfinWatchInfo | null {
-	if (!jellyfinMap || jellyfinMap.size === 0) return null;
-	const parsed = safeJsonParse(item.data);
-	if (!parsed) return null;
-	const data = parsed as Record<string, unknown>;
-	const remoteIds = data.remoteIds as Record<string, unknown> | undefined;
-	const tmdbId = remoteIds?.tmdbId;
-	if (!tmdbId) return null;
-	const mediaType = item.itemType === "movie" ? "movie" : "series";
-	return jellyfinMap.get(`${mediaType}:${tmdbId}`) ?? null;
-}
-
-function evaluateJellyfinLastWatched(
-	item: CacheItemForEval,
-	params: JellyfinLastWatchedParams,
-	ctx: EvalContext,
-): string | null {
-	const watch = lookupJellyfinWatch(item, ctx.jellyfinMap);
-	if (params.operator === "never") {
-		return !watch || watch.lastWatchedAt === null ? "Never watched (per Jellyfin)" : null;
-	}
-	if (!watch) return null;
-	if (watch.lastWatchedAt) {
-		const ageDays = (ctx.now.getTime() - watch.lastWatchedAt.getTime()) / (1000 * 60 * 60 * 24);
-		if (params.operator === "older_than" && params.days && ageDays >= params.days) {
-			return `Last watched ${Math.floor(ageDays)} days ago in Jellyfin (threshold: > ${params.days} days)`;
-		}
-		return null;
-	}
-	if (params.operator === "older_than" && params.days) {
-		const refDate = watch.addedAt ?? item.arrAddedAt;
-		if (refDate) {
-			const addedDays = (ctx.now.getTime() - refDate.getTime()) / (1000 * 60 * 60 * 24);
-			if (addedDays >= params.days) {
-				const source = watch.addedAt ? "Jellyfin" : "library";
-				return `Never watched, added to ${source} ${Math.floor(addedDays)} days ago (threshold: > ${params.days} days)`;
-			}
-		}
-	}
-	return null;
-}
-
-function evaluateJellyfinWatchCount(
-	item: CacheItemForEval,
-	params: JellyfinWatchCountParams,
-	ctx: EvalContext,
-): string | null {
-	const watch = lookupJellyfinWatch(item, ctx.jellyfinMap);
-	if (!watch) {
-		if (
-			ctx.jellyfinMap &&
-			ctx.jellyfinMap.size > 0 &&
-			params.operator === "less_than" &&
-			params.count > 0 &&
-			item.hasFile &&
-			item.arrAddedAt
-		) {
-			const ageDays = Math.floor(
-				(ctx.now.getTime() - item.arrAddedAt.getTime()) / (1000 * 60 * 60 * 24),
-			);
-			return `Not tracked by Jellyfin, in library for ${ageDays} days (threshold: < ${params.count} plays)`;
-		}
-		return null;
-	}
-	const count = watch.watchCount;
-	const ageCtx =
-		count === 0 && watch.addedAt
-			? `, added ${Math.floor((ctx.now.getTime() - watch.addedAt.getTime()) / (1000 * 60 * 60 * 24))} days ago`
-			: "";
-	if (params.operator === "less_than" && count < params.count)
-		return `Jellyfin play count: ${count}${ageCtx} (threshold: < ${params.count})`;
-	if (params.operator === "greater_than" && count > params.count)
-		return `Jellyfin play count: ${count} (threshold: > ${params.count})`;
-	return null;
-}
-
-function evaluateJellyfinOnDeck(
-	item: CacheItemForEval,
-	params: JellyfinOnDeckParams,
-	ctx: EvalContext,
-): string | null {
-	const watch = lookupJellyfinWatch(item, ctx.jellyfinMap);
-	if (!watch) return null;
-	if (params.isDeck && watch.onDeck) return "Item is on Jellyfin Continue Watching";
-	if (!params.isDeck && !watch.onDeck) return "Item is not on Jellyfin Continue Watching";
-	return null;
-}
-
-function evaluateJellyfinUserRating(
-	item: CacheItemForEval,
-	params: JellyfinUserRatingParams,
-	ctx: EvalContext,
-): string | null {
-	const watch = lookupJellyfinWatch(item, ctx.jellyfinMap);
-	if (params.operator === "unrated")
-		return !watch || watch.userRating === null ? "Unrated in Jellyfin" : null;
-	if (!watch || watch.userRating === null) return null;
-	if (
-		params.operator === "less_than" &&
-		params.rating !== undefined &&
-		watch.userRating < params.rating
-	)
-		return `Jellyfin rating: ${watch.userRating.toFixed(1)} (threshold: < ${params.rating})`;
-	if (
-		params.operator === "greater_than" &&
-		params.rating !== undefined &&
-		watch.userRating > params.rating
-	)
-		return `Jellyfin rating: ${watch.userRating.toFixed(1)} (threshold: > ${params.rating})`;
-	return null;
-}
-
-function evaluateJellyfinWatchedBy(
-	item: CacheItemForEval,
-	params: JellyfinWatchedByParams,
-	ctx: EvalContext,
-): string | null {
-	const watch = lookupJellyfinWatch(item, ctx.jellyfinMap);
-	if (!watch) return null;
-	const watchedBy = watch.watchedByUsers.map((u) => u.toLowerCase());
-	const targetNames = params.userNames.map((n) => n.toLowerCase());
-	if (params.operator === "includes_any") {
-		const matched = targetNames.filter((n) => watchedBy.includes(n));
-		if (matched.length > 0) return `Watched by Jellyfin user(s): ${matched.join(", ")}`;
-	} else if (params.operator === "excludes_all") {
-		if (targetNames.every((n) => !watchedBy.includes(n)))
-			return `Not watched by Jellyfin user(s): ${params.userNames.join(", ")}`;
-	}
-	return null;
-}
-
-function evaluateJellyfinAddedAt(
-	item: CacheItemForEval,
-	params: JellyfinAddedAtParams,
-	ctx: EvalContext,
-): string | null {
-	const watch = lookupJellyfinWatch(item, ctx.jellyfinMap);
-	if (!watch?.addedAt) return null;
-	const ageDays = (ctx.now.getTime() - watch.addedAt.getTime()) / (1000 * 60 * 60 * 24);
-	if (params.operator === "older_than" && ageDays >= params.days)
-		return `Added to Jellyfin ${Math.floor(ageDays)} days ago (threshold: > ${params.days} days)`;
-	if (params.operator === "newer_than" && ageDays < params.days)
-		return `Added to Jellyfin ${Math.floor(ageDays)} days ago (threshold: < ${params.days} days)`;
-	return null;
-}
-
-function evaluateJellyfinEpisodeCompletion(
-	item: CacheItemForEval,
-	params: JellyfinEpisodeCompletionParams,
-	ctx: EvalContext,
-): string | null {
-	if (item.itemType !== "series") return null;
-	const parsed = safeJsonParse(item.data);
-	if (!parsed) return null;
-	const data = parsed as Record<string, unknown>;
-	const tmdbId = (data.remoteIds as Record<string, unknown> | undefined)?.tmdbId;
-	if (typeof tmdbId !== "number") return null;
-
-	const stats = ctx.jellyfinEpisodeMap?.get(tmdbId);
-	if (!stats || stats.total === 0) return null;
-
-	let total: number;
-	let watched: number;
-	if (params.minSeason != null && stats.seasons.size > 0) {
-		total = 0;
-		watched = 0;
-		for (const [seasonNum, seasonStats] of stats.seasons) {
-			if (seasonNum >= params.minSeason) {
-				total += seasonStats.total;
-				watched += seasonStats.watched;
-			}
-		}
-		if (total === 0) return null;
-	} else {
-		total = stats.total;
-		watched = stats.watched;
-	}
-
-	const pct = (watched / total) * 100;
-	const seasonSuffix = params.minSeason != null ? ` (seasons >= ${params.minSeason})` : "";
-	if (params.operator === "less_than" && pct < params.percentage)
-		return `Jellyfin episode completion ${pct.toFixed(0)}% (${watched}/${total}) < ${params.percentage}%${seasonSuffix}`;
-	if (params.operator === "greater_than" && pct > params.percentage)
-		return `Jellyfin episode completion ${pct.toFixed(0)}% (${watched}/${total}) > ${params.percentage}%${seasonSuffix}`;
-	return null;
-}
-
 // ============================================================================
 // Phase C: New Rule Evaluators
 // ============================================================================
